+++ src/components/form/button-group/button-group.vue
@@ -4,7 +4,7 @@
 
 		<template #options="{ options, name }">
 			<slot name="options" v-bind="{ options, name }">
-				<div ref="optionsWrapperElement" class="mt-1 flex">
+				<div ref="optionsWrapperElement" class="flex mt-1">
 					<div v-for="option in options" :key="option.id">
 						<input ref="inputReferences" v-model="model" type="radio" class="peer sr-only" v-bind="{ id: option.id, value: option.value, name }" />